Fix bash linter errors

+++ init.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # This script configures the home directory to use shared config files
 # It creates symbolic links to all the files in this repo
 
@@ -7,19 +9,20 @@ ignore=('.' '..' '.git' '.gitignore' '.ssh' 'init.sh' 'README.md')
 # Create symbolic links to shared config files
 for filename in {.*,*}; do
     # Ignore names in Ignore
-    if [[ ! " ${ignore[@]} " =~ " $filename " ]]; then
-        echo Removing: "~/$filename"
-        rm -rf $HOME/$filename
-        echo Creating: "~/$filename" "-> ~/${PWD#$HOME/}/$filename"
-        ln -s $PWD/$filename $HOME/$filename
+    if [[ ! " ${ignore[@]} " =~ $filename ]]; then
+        echo "Removing: ~/$filename"
+        rm -rf "$HOME/${filename:?}"
+
+        echo Creating: "~/$filename -> ~/${PWD#$HOME/}/$filename"
+        ln -s "$PWD/$filename" "$HOME/$filename"
     else
-        echo Ignoring: $filename;
+        echo "Ignoring: $filename";
     fi
 done;
 
 # Install Vundle.vim and plugins
-rm -rf $HOME/.vim/bundle/Vundle.vim \
-    && mkdir -p $HOME/.vim/bundle \
-    && cd $HOME/.vim/bundle \
+rm -rf "$HOME/.vim/bundle/Vundle.vim" \
+    && mkdir -p "$HOME/.vim/bundle" \
+    && cd "$HOME/.vim/bundle" \
     && git clone https://github.com/VundleVim/Vundle.vim
 vim +PluginInstall +qall
